Moved CASounds to index to better fit existing project structure.
